requirements.txt
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
colored==1.4.4
idna==3.4
PySocks==1.7.1
PyYAML==6.0
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.2
